updated collections structue and add diagram
