+++ src/enums/role.go
@@ -1,0 +1,8 @@
+package enums
+
+type Role string
+
+const (
+	Admin    Role = "Admin"
+	Customer Role = "Customer"
+)
