--- src/role/role.go
@@ -1,1 +1,0 @@
-package role
