--- src/role/services.go
@@ -1,7 +1,0 @@
-package role
-
-type Role string
-
-const (
-	Admin Role = "Admin"
-)
