+++ src/router.go
@@ -3,8 +3,8 @@ package main
 import (
 	"go-api/src/auth"
 	"go-api/src/care"
+	"go-api/src/enums"
 	"go-api/src/middleware"
-	"go-api/src/role"
 	"go-api/src/utils"
 	"net/http"
 )
@@ -14,8 +14,8 @@ func loadRoutes(router *http.ServeMux, appData *utils.AppData) {
 	adminRouter := http.NewServeMux()
 	router.Handle("/auth/", http.StripPrefix("/auth", authRouter))
 
-	var adminAccess []role.Role = []role.Role{
-		role.Admin,
+	var adminAccess []enums.Role = []enums.Role{
+		enums.Admin,
 	}
 	router.Handle("/admin/", http.StripPrefix("/admin", middleware.Authenticate(adminAccess, appData)(adminRouter)))
 
@@ -27,5 +27,5 @@ func loadRoutes(router *http.ServeMux, appData *utils.AppData) {
 	adminRouter.HandleFunc("GET /Care/{id}", care.Get)
 	adminRouter.HandleFunc("POST /Care", care.Create)
 	adminRouter.HandleFunc("DELETE /Care/{id}", care.Delete)
-	
+
 }
